Remove id from /wards/schedule-confirmation
